refactor: moves airplane availability logic to DashboardPage
